+++ src/Bomberman/entity/Entity.java
@@ -5,8 +5,6 @@
  */
 package bomberman.entity;
 
-import javafx.scene.canvas.GraphicsContext;
-
 /**
  *
  * @author Ashish
