--- src/bomberman/GameState.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package bomberman;
-
-import java.util.Date;
-
-/**
- *
- * @author ashish.padalkar
- */
-public class GameState {
-    int level;
-    Date lastSaved;
-    boolean hasUnsavedChanges;
-}
