--- src/bomberman/constants/GlobalConstants.java
@@ -1,29 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package bomberman.constants;
-
-import javafx.scene.paint.Color;
-
-/**
- *
- * @author Ashish
- */
-public class GlobalConstants {
-
-    public static int sceneWidth = 500;
-    public static int sceneHeight = 500;
-    public static int canvasWidth = 500;
-    public static int canvasHeight = 500;
-    public static String GAME_NAME = "BomberMan";
-    public static String GAME_VERSION = "  v 0.1";
-    public static Color backGroundColor = Color.WHITE;
-    public static int playerX = 200;
-    public static int playerY = 200;
-    public static int playerWidth = 18;
-    public static int playerHeight = 21;
-    public static int playerScale = 2;
-
-}
